src/analyzers/overpermission_analyzer/requirements.txt
# 共通の依存関係を読み込む
-r ../../../common_requirements.txt

# このFunction固有のライブラリ
google-cloud-recommender
